# api_gateway/OpenAPI.yml
openapi: 3.0.3
info:
  title: StudyFlow API
  version: 1.0.0
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
...
    Slot:
      type: object
      properties:
        id:
          type: string
        tutorId:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        isBooked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    Lesson:
      type: object
      properties:
        id:
          type: string
        slotId:
          type: string
        studentId:
          type: string
        status:
          $ref: '#/components/schemas/LessonStatus'
        isPaid:
          type: boolean
        connectionLink:
          type: string
        priceRub:
          type: integer
        paymentInfo:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    LessonStatus:
      type: string
      enum:
        - booked
        - cancelled
        - completed
...
paths:
...
  /schedule/slots:
    post:
      summary: Create a slot
      operationId: createSlot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tutorId:
                  type: string
                startsAt:
                  type: string
                  format: date-time
                endsAt:
                  type: string
                  format: date-time
              required:
                - tutor_id
                - starts_at
                - ends_at
      responses:
        '200':
          description: Slot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Slot'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedule/slots/{id}:
    get:
      summary: Get a slot
      operationId: getSlot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Slot info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Slot'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update a slot
      operationId: updateSlot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startsAt:
                  type: string
                  format: date-time
                endsAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Slot updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Slot'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Slot already booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a slot
      operationId: deleteSlot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Slot deleted
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Slot is booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedule/slots/by-tutor/{tutor_id}:
    get:
      summary: List slots by tutor
      operationId: listSlotsByTutor
      parameters:
        - name: tutor_id
          in: path
          required: true
          schema:
            type: string
        - name: only_available
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedule/lessons:
    get:
      summary: List lessons
      operationId: listLessons
      parameters:
        - name: tutor_id
          in: query
          schema:
            type: string
        - name: student_id
          in: query
          schema:
            type: string
        - name: status_filter
          in: query
          schema:
            $ref: '#/components/schemas/LessonStatus'
          explode: true
          style: form
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid combination of query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a lesson
      operationId: createLesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                slotId:
                  type: string
                studentId:
                  type: string
              required:
                - slot_id
                - student_id
      responses:
        '200':
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          description: Invalid argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Slot already booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Tutor and student are not linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedule/lessons/{id}:
    get:
      summary: Get a lesson
      operationId: getLesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lesson info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update a lesson
      operationId: updateLesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                connectionLink:
                  type: string
                priceRub:
                  type: integer
                paymentInfo:
                  type: string
      responses:
        '200':
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedule/lessons/{id}/cancel:
    post:
      summary: Cancel a lesson
      operationId: cancelLesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lesson cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


  # payment
...